pyproject.toml
[project]
name = "disona"
version = "0.1.0"
description = "A discord sona engine based on gemini flash"
authors = [
    {name = "Danon",email = "danyapancresnicov@gmail.com"}
]
license = {text = "GPLv3"}
readme = "README.md"
requires-python = "3.10.16"
dependencies = [
    "discord (>=2.3.2,<3.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "toml (>=0.10.2,<0.11.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
